+++ src/core/pxtnDelay.cpp
@@ -1,15 +1,17 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "./pxTypedef.h"
+
 #include "./pxtnMax.h"
 
 #include "./pxtnDelay.h"
 
 pxtnDelay::pxtnDelay()
 {
-	for( int i = 0; i < MAX_CHANNEL; i ++ ) _bufs[ i ] = NULL;
+	for( s32 i = 0; i < MAX_CHANNEL; i ++ ) _bufs[ i ] = NULL;
 	_smp_num   =    0;
-	_b_played = true;
+	_b_played  = true;
 }
 
 pxtnDelay::~pxtnDelay()
@@ -23,11 +25,11 @@ bool pxtnDelay::Init()
 }
 */
 DELAYUNIT pxtnDelay::get_unit ()const { return _unit ; }
-int       pxtnDelay::get_group()const { return _group; }
-float     pxtnDelay::get_rate ()const { return _rate ; }
-float     pxtnDelay::get_freq ()const { return _freq ; }
+s32       pxtnDelay::get_group()const { return _group; }
+f32       pxtnDelay::get_rate ()const { return _rate ; }
+f32       pxtnDelay::get_freq ()const { return _freq ; }
 
-void      pxtnDelay::Set( DELAYUNIT unit, float freq, float rate, int group )
+void      pxtnDelay::Set( DELAYUNIT unit, f32 freq, f32 rate, s32 group )
 {
 	_unit  = unit ;
 	_group = group;
@@ -43,11 +45,11 @@ bool pxtnDelay::switch_played(){ _b_played = _b_played ? false : true; return _b
 
 void pxtnDelay::Tone_Release()
 {
-	for( int i = 0; i < MAX_CHANNEL; i ++ ){ if( _bufs[i] ) free( _bufs[i] ); _bufs[i] = NULL; }
+	for( s32 i = 0; i < MAX_CHANNEL; i ++ ){ if( _bufs[i] ) free( _bufs[i] ); _bufs[i] = NULL; }
 	_smp_num = 0;
 }
 
-bool pxtnDelay::Tone_Ready( int beat_num, float beat_tempo, int sps, int bps )
+bool pxtnDelay::Tone_Ready( s32 beat_num, f32 beat_tempo, s32 sps, s32 bps )
 {
 	Tone_Release();
 
@@ -56,19 +58,19 @@ bool pxtnDelay::Tone_Ready( int beat_num, float beat_tempo, int sps, int bps )
 	bool  b_ret = false;
 
 	_offset    = 0;
-	_rate_long = (long)_rate;// / 100;
+	_rate_long = (s32)_rate;// / 100;
 
 	switch( _unit )
 	{
-	case DELAYUNIT_Beat  : _smp_num = (int)( sps * 60            / beat_tempo / _freq ); break;
-	case DELAYUNIT_Meas  : _smp_num = (int)( sps * 60 * beat_num / beat_tempo / _freq ); break;
-	case DELAYUNIT_Second: _smp_num = (int)( sps                              / _freq ); break;
+	case DELAYUNIT_Beat  : _smp_num = (s32)( sps * 60            / beat_tempo / _freq ); break;
+	case DELAYUNIT_Meas  : _smp_num = (s32)( sps * 60 * beat_num / beat_tempo / _freq ); break;
+	case DELAYUNIT_Second: _smp_num = (s32)( sps                              / _freq ); break;
 	}
 
-	for( int c = 0; c < MAX_CHANNEL; c++ )
+	for( s32 c = 0; c < MAX_CHANNEL; c++ )
 	{
-		if( !(  _bufs[ c ] = (int*)malloc( _smp_num * sizeof(int) ) ) ) goto End;
-		memset( _bufs[ c ], 0,             _smp_num * sizeof(int) );
+		if( !(  _bufs[ c ] = (s32*)malloc( _smp_num * sizeof(s32) ) ) ) goto End;
+		memset( _bufs[ c ], 0,             _smp_num * sizeof(s32) );
 	}
 
 	b_ret = true;
@@ -79,10 +81,10 @@ End:
 	return b_ret;
 }
 
-void pxtnDelay::Tone_Supple( int ch, long *group_smps )
+void pxtnDelay::Tone_Supple( s32 ch, s32 *group_smps )
 {
 	if( !_smp_num ) return;
-	int a = _bufs[ ch ][ _offset ] * _rate_long / 100;
+	s32 a = _bufs[ ch ][ _offset ] * _rate_long / 100;
 	if( _b_played ) group_smps[ _group ] += a;
 	_bufs[ ch ][ _offset ] =  group_smps[ _group ];
 }
@@ -93,43 +95,43 @@ void pxtnDelay::Tone_Increment()
 	if( ++_offset >= _smp_num ) _offset = 0;
 }
 
-void  pxtnDelay::Tone_Clear( void )
+void pxtnDelay::Tone_Clear( void )
 {
 	if( !_smp_num ) return;
 
-	int def = 0; // ..
+	s32 def = 0; // ..
 
 	//	TuneData_Quality_Get( NULL, NULL, &bps );
 
-	for( int i = 0; i < MAX_CHANNEL; i ++ ) memset( _bufs[ i ], def, _smp_num * sizeof(int) );
+	for( s32 i = 0; i < MAX_CHANNEL; i ++ ) memset( _bufs[ i ], def, _smp_num * sizeof(s32) );
 }
 
 
 // (12byte) =================
 typedef struct
 {
-	unsigned short unit ;
-	unsigned short group;
-	float          rate ;
-	float          freq ;
+	u16 unit ;
+	u16 group;
+	f32 rate ;
+	f32 freq ;
 }
 _DELAYSTRUCT;
 
 bool pxtnDelay::Write( pxwrDoc *p_doc ) const
 {
-	_DELAYSTRUCT    dela;
-	long            size;
+	_DELAYSTRUCT dela;
+	s32          size;
 
 	memset( &dela, 0, sizeof( _DELAYSTRUCT ) );
-	dela.unit  = (unsigned short)_unit ;
-	dela.group = (unsigned short)_group;
+	dela.unit  = (u16)_unit ;
+	dela.group = (u16)_group;
 	dela.rate  = _rate;
 	dela.freq  = _freq;
 
 	// dela ----------
 	size = sizeof( _DELAYSTRUCT );
-	if( !p_doc->w( &size, sizeof(unsigned long), 1 ) ) return false;
-	if( !p_doc->w( &dela, size,                  1 ) ) return false;
+	if( !p_doc->w( &size, sizeof(u32), 1 ) ) return false;
+	if( !p_doc->w( &dela, size,        1 ) ) return false;
 
 	return true;
 }
@@ -137,7 +139,7 @@ bool pxtnDelay::Write( pxwrDoc *p_doc ) const
 bool pxtnDelay::Read( pxwrDoc *p_doc, bool *pb_new_fmt )
 {
 	_DELAYSTRUCT dela;
-	long         size;
+	s32          size;
 
 	memset( &dela, 0, sizeof(_DELAYSTRUCT) );
 	if( !p_doc->r( &size, 4,                    1 ) ) return false;
